Build server via dist as well
