+++ xtask/src/main.rs
@@ -13,10 +13,11 @@ use std::env;
 use pico_args::Arguments;
 use xtask::{
     codegen::{self, Mode},
+    dist::run_dist,
     install::{ClientOpt, InstallCmd, ServerOpt},
     not_bash::pushd,
-    pre_commit, project_root, run_clippy, run_dist, run_fuzzer, run_pre_cache, run_release,
-    run_rustfmt, Result,
+    pre_commit, project_root, run_clippy, run_fuzzer, run_pre_cache, run_release, run_rustfmt,
+    Result,
 };
 
 fn main() -> Result<()> {
